+++ src/index.html
@@ -93,8 +93,8 @@
             Spooky
           </vr-text>  
         </vr-div>
-        <vr-div id="column-div-1" background-color="#602202" height="100" margin="20" layout="Column" padding="20">
-          <vr-div id="demo-div-1" margin="20" background-color="#222222" padding="20">
+        <vr-div id="column-div-1" background-color="#602202" height="100" margin="20" layout="Column" padding="20" item-vertical-align="Top">
+          <vr-div  width="500" id="demo-div-1"  margin="20" background-color="#222222" padding="20">
             <vr-text id="360-video-button" border-radius="10" height="400" font-color="#222222" background-color="#ffffff" padding="15">
               Play 360 Video
             </vr-text>
@@ -123,7 +123,7 @@
               this is it
             </vr-text>
           </vr-div>
-          <vr-div id="demo-div-2" background-color="#000fff" margin="5" padding="10" height="800" vertical-align="Middle">
+          <vr-div id="demo-div-2" height="600" width="500" background-color="#ff0000" margin="5" vertical-align="Middle">
             <vr-text width="300" font-color="#222222" background-color="#ffffff" padding="10">
               not now
 
@@ -175,7 +175,7 @@
         
         // ****** Testing divs
 
-        setTimeout(() => {
+        /*setTimeout(() => {
           const demoDiv1 = document.getElementById("demo-div-1");
           demoDiv1.backgroundColor = "#ff0000";
           demoDiv1.borderRadius = 20;
@@ -198,7 +198,7 @@
               demoDiv2.horizontalAlign = "Right";
             }, 3000);
           }, 3000);
-        }, 3000);
+        }, 3000);*/
 
         // ****** Testing images
